+++ app/Filament/Resources/AssessmentPeriodResource.php
@@ -42,7 +42,7 @@ class AssessmentPeriodResource extends Resource
 
     protected static string|\BackedEnum|null $navigationIcon = 'heroicon-o-calendar-days';
 
-    protected static string|\UnitEnum|null $navigationGroup = 'Penilaian';
+    protected static string|\UnitEnum|null $navigationGroup = 'Manajemen Sekolah';
 
     protected static ?string $navigationLabel = 'Periode Penilaian';