+++ app/Filament/Resources/AssessmentSchemeResource.php
@@ -33,7 +33,7 @@ class AssessmentSchemeResource extends Resource
 
     protected static string|\BackedEnum|null $navigationIcon = 'heroicon-o-adjustments-horizontal';
 
-    protected static string|\UnitEnum|null $navigationGroup = 'Penilaian';
+    protected static string|\UnitEnum|null $navigationGroup = 'Manajemen Sekolah';
 
     protected static ?string $navigationLabel = 'Komponen dan Bobot';